+++ src/main/java/roomescape/reservation/controller/ReservationApiController.java
@@ -7,7 +7,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -19,7 +18,6 @@ import roomescape.reservation.dto.request.ReservationSaveRequest;
 import roomescape.reservation.dto.request.ReservationSearchCondRequest;
 import roomescape.reservation.dto.response.MemberReservationResponse;
 import roomescape.reservation.dto.response.ReservationResponse;
-import roomescape.reservation.dto.response.WaitingResponse;
 import roomescape.reservation.service.ReservationService;
 
 @RestController
@@ -74,38 +72,9 @@ public class ReservationApiController {
                 .body(reservationResponse);
     }
 
-    @GetMapping("/reservations/wait")
-    public ResponseEntity<ResourcesResponse<WaitingResponse>> findWaitingReservations() {
-        List<WaitingResponse> reservations = reservationService.findWaitingReservations();
-        ResourcesResponse<WaitingResponse> response = new ResourcesResponse<>(reservations);
-        return ResponseEntity.ok(response);
-    }
-
-    @PostMapping("/reservations/wait")
-    public ResponseEntity<ReservationResponse> createMemberWaitingReservation(
-            @Valid @RequestBody ReservationSaveRequest reservationSaveRequest,
-            LoginMember loginMember
-    ) {
-        ReservationResponse reservationResponse = reservationService.save(
-                reservationSaveRequest,
-                loginMember,
-                Status.WAIT
-        );
-
-        return ResponseEntity.created(URI.create("/reservations/wait/" + reservationResponse.id()))
-                .body(reservationResponse);
-    }
-
     @DeleteMapping("/reservations/{id}")
     public ResponseEntity<Void> delete(@PathVariable("id") Long id) {
         reservationService.delete(id);
-
-        return ResponseEntity.noContent().build();
-    }
-
-    @PatchMapping("/reservations/wait/{id}")
-    public ResponseEntity<Void> updateSuccess(@PathVariable("id") Long id) {
-        reservationService.updateSuccess(id);
 
         return ResponseEntity.noContent().build();
     }
